Rename to helpers, add func
