+++ mev_inspect/classifiers/helpers.py
@@ -84,3 +84,14 @@ def _filter_transfers(
         filtered_transfers.append(transfer)
 
     return filtered_transfers
+
+
+def get_amount_transferred_to_address(
+    address: str, transfers: List[Transfer]
+) -> Optional[int]:
+
+    for transfer in transfers:
+        if transfer.to_address == address:
+            return transfer.amount
+
+    raise ValueError(f"Transfer to {address} not found.")
